backend/requirements.txt
fastapi>=0.115
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
alembic
arq
redis
sse-starlette
python-jose[cryptography]
passlib[bcrypt]
boto3
anthropic>=0.40.0
networkx>=3.0
pyyaml
python-dotenv
pydantic-settings
python-multipart
email-validator
